+++ src/vst/pst.hpp
@@ -1,0 +1,7 @@
+#include "../graph/spanning_tree.hpp"
+#include "../graph/tree.hpp"
+
+namespace pst {
+
+tree::Tree compute_pst(spanning_tree::Tree &st);
+} // namespace pst
