+++ src/vst/vst.cpp
@@ -1,6 +1,3 @@
-#include "./vst.hpp"
-#include "../graph/graph.hpp"
-#include "../graph/spanning_tree.hpp"
 #include <algorithm>
 #include <cstddef>
 #include <cstdint>
@@ -13,6 +10,9 @@
 #include <stack>
 
 #include "./vst.hpp"
+#include "../graph/tree.hpp"
+#include "../graph/graph.hpp"
+#include "../graph/spanning_tree.hpp"
 
 namespace vst {
 /*
@@ -324,271 +324,5 @@ void cycle_equiv(spanning_tree::Tree &t) {
 
 } // namespace vst
 
-namespace tree {
-  // TODO: declare all of these in a header file
-  const std::size_t SIZE_T_MAX = std::numeric_limits<size_t>::max();
-
-  // Vertex
-  // ======
-  
-  // constructor(s)
-  
-  Vertex::Vertex() :
-    id(SIZE_T_MAX), parent(SIZE_T_MAX), children(std::set<std::size_t>{}), is_valid_(false) {}
-  Vertex::Vertex(std::size_t id) :
-    id(id), parent(SIZE_T_MAX), children(std::set<std::size_t>{}), is_valid_(true) {}
-  Vertex::Vertex(std::size_t id, std::size_t parent_id) :
-    id(id), parent(parent_id), children(std::set<std::size_t>{}), is_valid_(true) {}
-
-  // member function(s)
-  // ------------------
-
-  // getters
-  bool Vertex::is_valid() const { return this->is_valid_; }
-  std::set<std::size_t> const& Vertex::get_children() const { return this->children; }
-
-  // setters
-  // -------
-  
-  void Vertex::add_child(std::size_t child_id) {
-    this->children.insert(child_id);
-  }
-  
-  // Tree
-  // ====
-
-  // constructor(s)
-  
-  Tree::Tree() : vertices(std::vector<tree::Vertex>{}) {}
-
-  Tree::Tree(std::size_t n) : vertices(std::vector<Vertex>(n, Vertex())) {}
-
-  // member function(s)
-  // ------------------
-
-  // setters
-  bool Tree::add_vertex(std::size_t parent_id, std::size_t id) {
-    // TODO: there's a logical error in the caller if the vertex is already in the tree
-    //       should we throw an exception here?
-    if (this->vertices[id].is_valid()) { return false;  }
-    this->vertices[id] = Vertex(id, parent_id);
-    this->vertices[parent_id].add_child(id);
-    return true;
-  }
-
-  std::set<std::size_t> const& Tree::get_children(std::size_t v) const {
-    return this->vertices.at(v).get_children();
-  }
-
-  void Tree::print_dot() {
-  std::cout << std::format(
-    "graph G {{\n"
-    "\trankdir = TB;\n"
-    "\tnode[shape = circle];\n"
-    "\tedge [arrowhead=vee];\n"
-  );
-
-  for (std::size_t i{}; i < this->size(); i++) {
-    for (auto c : this->get_children(i)) {
-      std::cout << std::format("\t{} -- {};\n", i, c);
-    }
-  }
-
-  std::cout << "}" << std::endl;
-}
-
-  // TODO: should this be the number of valid vertices?
-  std::size_t Tree::size() const { return this->vertices.size(); }
-  
-} // namespace tree
-
-
-namespace pst {
-
-// constructor(s)
-// --------------
-PST::PST(spanning_tree::Tree &t) : t(std::vector<vst::Vertex>{}) {
-  this->t.reserve(t.size()); //
-  // for DFS
-  std::stack<std::size_t> vertex_stack{}; // stack of vertices
-
-  // to determine out parent child relationships in the PST
-  std::stack<std::size_t> class_stack{}; // stack of classes
-
-  // to keep track of which vertices we pushed onto the vertex stack (seen)
-  std::set<std::size_t> seen{};
-
-  // a mapping of a given class and its parent class
-  std::map <std::size_t, std::size_t> class_and_parent{};
-
-  // initialize the root class parent to itself (0)
-  class_and_parent[0] = 0;
-
-  std::size_t current_class{};
-  class_stack.push(current_class);
-
-  std::size_t current_vertex{};
-  vertex_stack.push(current_vertex);
-
-  // initialize the PST with the root vertex
-  vst::Vertex root =
-    vst::Vertex{current_vertex, current_class};
-  this->t.push_back(root);
-
-  vst::Vertex v = root;
-
-  while (!vertex_stack.empty()) {
-    current_vertex = vertex_stack.top();
-    seen.insert(current_vertex);
-
-    if (current_vertex > 0) {
-      spanning_tree::Edge& current_edge = t.get_incoming_edge(current_vertex);
-      std::size_t new_class = current_edge.get_class_idx();
-
-      auto it = class_and_parent.find(new_class);
-      if (it == class_and_parent.end()) {
-        std::size_t cl_idx = this->t.size();
-        class_and_parent[new_class] = cl_idx;
-        //class_and_parent.find(current_class)->second;
-
-        vst::Vertex v = vst::Vertex{current_vertex,
-                                    current_class,
-                                    class_and_parent.find(current_class)->second
-                                   };
-
-        this->t.push_back(v);
-
-        if (current_class < this->t.size())   {
-          this->t[current_class].add_child(cl_idx);
-        } else {
-          std::cout << "could not add current_vertex: " << current_vertex << "\n";
-        }
-      }
-
-      current_class = new_class;
-    }
-
-    bool explored = true;
-
-    for (auto c: t.get_children(current_vertex)){
-      if (seen.find(c) == seen.end()) {
-        vertex_stack.push(c);
-        explored = false;
-      }
-    }
-
-    if (explored) {
-      vertex_stack.pop();
-    }
-  }
-}
-
-std::set<std::size_t> const& PST::get_children(std::size_t v) const {
-  return this->t.at(v).get_children();
-}
-
-std::size_t PST::size() const {
-  return this->t.size();
-}
-
-void PST::print_dot() {
-  std::cout << std::format(
-    "graph G {{\n"
-    "\trankdir = TB;\n"
-    "\tnode[shape = circle];\n"
-    "\tedge [arrowhead=vee];\n"
-  );
-
-  for (std::size_t i{}; i < this->size(); i++) {
-    for (auto c : this->get_children(i)) {
-      std::cout << std::format("\t{} -- {};\n", i, c);
-    }
-  }
-
-  std::cout << "}" << std::endl;
-}
-
-/*
- * PST
- * ---
- *
- * The PST is a tree of cycle equivalent regions of the spanning tree of the
- * undirected flow graph U_f with the following properties:
-  * -
-  *
-
-  we use region to refer to a cycle equivalent region of the spanning tree
-  therefore region is synonymous with class
-  
- * The construction is as follows:
-     "When a region is first entered, we set its parent to the current region 
-     and then update the current region to be the region just entered.
-     When a region is exited,
-     the current region is set to be the exited region’s parent."
-  
-  * When a region is first entered:
-      We know a region has been entered because: 
-        - there's an OBE in the current vertex (capping or not)
-    1. we set its parent to the current region
-       and then update the current region to be the region just entered
-       * its parent will be the class on top of the stack
-         * if the stack was empty this class will be the root
-    2. we will then push it onto the stack making it the current region
-       the current equiv class is the class on top of the stack
-   
-  * When a region is exited:
-      We know a region has been entered because: 
-       - we see an outgoing capping back-edge
-       - we see an incoming back-edge
-    1 the current region is set to be the exited region’s parent.
-      a) we pop the value that's on top of the stack
-      b) st spanning tree  
- */
-tree::Tree compute_pst(spanning_tree::Tree &st) {
-  tree::Tree t = tree::Tree(st.size());
-  std::stack<std::size_t> vertex_stack{}; // stack of vertices
-
-  // go over the nodes in reverse topological order (dfs_num) which may not be
-  // equal to node ids
-  for (std::size_t r{st.size() - 1}; r > 0; --r) {
-    // the vertex id of the vertex at the topological sort index r
-    std::size_t v = st.get_sorted(r);
-    
-    std::set<size_t> obe = st.get_obe_idxs(v);
-    std::set<size_t> ibe = st.get_ibe(v);
-
-    if (obe.empty() && ibe.empty()) { continue; }
-    
-    // the equivalence class of the incoming tree edge
-    std::size_t cl = st.get_parent_edge(v).get_class();
-
-    // check if any of the obe are capping backedges
-    bool has_capping_obe{false};
-    for (auto e_idx: obe) {
-      if (st.get_backedge(e_idx).is_capping_backedge()) {
-        has_capping_obe=true;
-        break;
-      }
-    }
-
-    // exit the current region
-    // --------------------
-    
-    if (!ibe.empty() || has_capping_obe) {
-      vertex_stack.pop();
-    }
-
-    // enter the new region
-    // --------------------
-
-    if (!obe.empty()) { // an equivalence class/region has been entered
-      if (!vertex_stack.empty()) { t.add_vertex(vertex_stack.top(), cl); }
-      vertex_stack.push(cl);
-    }
-  }
-
-  return t;
-}
-} // namespace pst
